use should for expectations
I think boilerplate is great for creating apps fast.
This has not panned out very well because I am not familiar with jest's
mocking/stubbing mechanisms. At that point, I chose to mock/stub with sinon
which is what I've used in the past. That pushed me down the path of using
chai/sinon-chai (chai.should()) for assertions.